14th Session
I fixed the issue of the train carriages starting transparent by removing all but the player's rigidbody. I added in money which gets deleted when the player goes up to it and it adds to the "moneyCollected" variable. I also added in functionality for the train to go at different speeds, which works almost flawlessly except that when going from high to low speed there is a small gap in the terrain. I also need to add the guards still.